pyproject.toml
[project]
name = "a"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.3",
    "groq>=1.1.2",
    "httpx>=0.28.1",
    "joblib>=1.5.3",
    "numpay>=0.0.8",
    "numpy>=2.4.4",
    "pandas>=3.0.2",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
    "scikit-learn==1.6.1",
    "streamlit>=1.56.0",
    "uvicorn>=0.44.0",
    "xgboost>=3.2.0",
]
